Continue splitting Level and World, fix state
